+++ src/UI/SettingsWindow.py
@@ -8,8 +8,6 @@ currentdir = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentfram
 parentdir = os.path.dirname(currentdir)
 sys.path.insert(0,parentdir)
 
-from Utils.Utils import Actions, Markets
-from Utils.Trade import Trade
 from .WarningWindow import WarningWindow
 
 class SettingsWindow(tk.Toplevel):