+++ CHANGELOG.md
@@ -8,7 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Issue37 - Improved installation process and dependencies setup
 - Updated default .credentials configured path
+
+### Added
 - Added Pipfile to manage python dependencies
+- Added `FEE` action
+- Added `notes` field in trade
 
 ## [1.0.0] 2019-05-03
 ### Added